lint & format
